+++ packages/localizations/src/esCo.ts
@@ -109,6 +109,7 @@ export const esCo = {
       offline: 'Pagos sin conexión',
       ach: 'Cuenta bancaria',
       mercadopago: 'Mercado Pago',
+      ccavenue: 'Pagar con CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -120,6 +121,7 @@ export const esCo = {
       ach: '',
       mercadopago:
         'Usa el formulario de MercadoPago a continuación para completar tu compra de forma segura.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'No hay métodos de pago disponibles',
     cardNumber: 'Número de tarjeta',