+++ packages/localizations/src/frFr.ts
@@ -109,6 +109,7 @@ export const frFr = {
       offline: 'Paiements hors ligne',
       ach: 'Compte bancaire',
       mercadopago: 'Mercado Pago',
+      ccavenue: 'Payer avec CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -120,6 +121,7 @@ export const frFr = {
       ach: '',
       mercadopago:
         'Utilisez le formulaire MercadoPago ci-dessous pour finaliser votre achat en toute sécurité.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'Aucune méthode de paiement disponible',
     cardNumber: 'Numéro de carte',