+++ packages/localizations/src/itIt.ts
@@ -109,6 +109,7 @@ export const itIt = {
       offline: 'Pagamenti offline',
       ach: 'Conto Bancario',
       mercadopago: 'Mercado Pago',
+      ccavenue: 'Paga con CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -120,6 +121,7 @@ export const itIt = {
       ach: '',
       mercadopago:
         'Usa il modulo MercadoPago qui sotto per completare l’acquisto in modo sicuro.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'Nessun metodo di pagamento disponibile',
     cardNumber: 'Numero della carta',