+++ packages/react/src/components/checkout/payment/lazy-payment-loader.tsx
@@ -138,6 +138,11 @@ const LazyComponents = {
       module => ({ default: module.PazeCheckoutButton })
     )
   ),
+  CCAvenueCheckoutButton: lazy(() =>
+    import(
+      '@/components/checkout/payment/checkout-buttons/ccavenue/ccavenue'
+    ).then(module => ({ default: module.CCAvenueCheckoutButton }))
+  ),
 
   // Container Components
   CreditCardContainer: lazy(() =>
@@ -214,6 +219,11 @@ type PaymentComponentRegistry = {
       button: PaymentComponentKey;
     };
   };
+  [PaymentMethodType.CCAVENUE]?: {
+    [PaymentProvider.CCAVENUE]: {
+      button: PaymentComponentKey;
+    };
+  };
 };
 
 export const lazyPaymentComponentRegistry: PaymentComponentRegistry = {
@@ -272,6 +282,11 @@ export const lazyPaymentComponentRegistry: PaymentComponentRegistry = {
     [PaymentProvider.GODADDY]: {
       form: 'GoDaddyACHForm',
       button: 'ACHCheckoutButton',
+    },
+  },
+  [PaymentMethodType.CCAVENUE]: {
+    [PaymentProvider.CCAVENUE]: {
+      button: 'CCAvenueCheckoutButton',
     },
   },
 };