+++ packages/react/src/components/checkout/payment/payment-form.tsx
@@ -23,6 +23,7 @@ import {
   type Product,
 } from '@/components/checkout/line-items';
 import ApplePayIcon from '@/components/checkout/payment/icons/ApplePay';
+import CcavenueIcon from '@/components/checkout/payment/icons/Ccavenue';
 import GooglePayIcon from '@/components/checkout/payment/icons/GooglePay';
 import MercadoPagoIcon from '@/components/checkout/payment/icons/MercadoPago';
 import PayPalIcon from '@/components/checkout/payment/icons/PayPal';
@@ -76,6 +77,7 @@ const PAYMENT_METHOD_ICONS: Record<string, React.ReactNode> = {
   paze: <PazeIcon className='h-5 w-8' />,
   mercadopago: <MercadoPagoIcon className='h-5 w-8' />,
   offline: <Wallet className='h-5 w-5' />,
+  ccavenue: <CcavenueIcon className='h-5 w-5' />,
 };
 
 export function PaymentForm(
@@ -127,6 +129,8 @@ export function PaymentForm(
           return t.payment.methods.offline;
         case PaymentMethodType.MERCADOPAGO:
           return t.payment.methods.mercadopago;
+        case PaymentMethodType.CCAVENUE:
+          return t.payment.methods.ccavenue;
         default:
           return key;
       }
@@ -154,6 +158,8 @@ export function PaymentForm(
           return t.payment.descriptions?.offline;
         case PaymentMethodType.MERCADOPAGO:
           return t.payment.descriptions?.mercadopago;
+        case PaymentMethodType.CCAVENUE:
+          return t.payment.descriptions?.ccavenue;
         default:
           return undefined;
       }