+++ packages/react/src/types.ts
@@ -44,6 +44,7 @@ export const PaymentProvider = {
   PAZE: 'paze',
   OFFLINE: 'offline',
   MERCADOPAGO: 'mercadopago',
+  CCAVENUE: 'ccavenue',
 } as const;
 
 export const CheckoutType = {
@@ -65,6 +66,7 @@ export const PaymentMethodType = {
   OFFLINE: 'offline',
   PAZE: 'paze',
   MERCADOPAGO: 'mercadopago',
+  CCAVENUE: 'ccavenue',
 } as const;
 
 // Union of all payment method keys